+++ tests/Data/Model/Simple.php
@@ -22,7 +22,7 @@ class Simple extends Model
 
         $this->addField($this->fieldName()->x, ['type' => 'string', 'required' => true]);
 
-        $this->addField($this->fieldName()->refId, ['type' => 'integer']);
+        $this->addField($this->fieldName()->refId, ['type' => 'bigint']);
         $this->hasOne($this->fieldName()->ref, ['model' => [Standard::class], 'ourField' => $this->fieldName()->refId]);
     }
 }
